requirements.txt
opencv-python
pyautogui
pytesseract
Pillow
easyocr